Refactored connector/httpParser to fix scoverage compilation issue
