+++ build.sbt
@@ -29,14 +29,14 @@ lazy val coverageSettings: Seq[Setting[_]] = {
 
   val excludedPackages = Seq(
     "<empty>",
-    "Reverse.*",
+    ".*Reverse.*",
     ".*standardError*.*",
     "uk.gov.hmrc.BuildInfo",
     "app.*",
     "prod.*",
-    "uk.gov.hmrc.vatsubscription.config.*",
+    "config.*",
     "testOnlyDoNotUseInAppConf.*",
-    "uk.gov.hmrc.vatsubscription.testonly.*"
+    "testonly.*"
   )
 
   Seq(
