+++ test/httpparsers/GetVatCustomerInformationHttpParserSpec.scala
@@ -14,24 +14,24 @@
  * limitations under the License.
  */
 
-package connectors
+package httpparsers
 
 import assets.TestUtil
-import play.api.http.Status.{BAD_REQUEST, FORBIDDEN, INTERNAL_SERVER_ERROR, NOT_FOUND, OK}
+import config.featureSwitch.Api1363Latest
+import connectors._
+import helpers.CustomerInformationTestConstants._
+import play.api.http.Status._
 import play.api.libs.json.Json
 import uk.gov.hmrc.http.HttpResponse
-import uk.gov.hmrc.http.HttpClient
-import config.featureSwitch.Api1363Latest
-import helpers.CustomerInformationTestConstants._
 
-class GetVatCustomerInformationConnectorSpec extends TestUtil {
+class GetVatCustomerInformationHttpParserSpec extends TestUtil {
 
   val testHttpVerb = "GET"
   val testUri = "/"
 
-  "GetVatCustomerInformationConnector GetVatCustomerInformationHttpReads" should {
+  "GetVatCustomerInformationHttpReads" should {
 
-    val httpClient = mock[HttpClient]
+    val httpParser = new GetVatCustomerInformationHttpParser(mockAppConfig)
 
     "read" should {
 
@@ -39,77 +39,63 @@ class GetVatCustomerInformationConnectorSpec extends TestUtil {
 
         mockAppConfig.features.api1363Version(Api1363Latest)
 
-        val mockConnector: GetVatCustomerInformationConnector = new GetVatCustomerInformationConnector(httpClient, mockAppConfig)
-
         val httpResponse = HttpResponse(OK, customerInformationDESJsonMaxWithFRS.toString)
 
-        val res = mockConnector.GetVatCustomerInformationHttpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
+        val res = httpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
 
         res shouldBe Right(customerInformationModelMaxWithFRS)
       }
 
       "parse an OK response with an invalid json as a UnexpectedGetVatCustomerInformationFailure" in {
 
-        val mockConnector: GetVatCustomerInformationConnector = new GetVatCustomerInformationConnector(httpClient, mockAppConfig)
-
         val httpResponse = HttpResponse(OK,Json.obj().toString)
 
-        val res = mockConnector.GetVatCustomerInformationHttpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
+        val res = httpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
 
         res shouldBe Left(UnexpectedGetVatCustomerInformationFailure(INTERNAL_SERVER_ERROR, "Invalid Success Response Json"))
       }
 
       "parse a BAD_REQUEST response as a InvalidVatNumber" in {
 
-        val mockConnector: GetVatCustomerInformationConnector = new GetVatCustomerInformationConnector(httpClient, mockAppConfig)
-
         val httpResponse = HttpResponse(BAD_REQUEST, "")
 
-        val res = mockConnector.GetVatCustomerInformationHttpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
+        val res = httpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
 
         res shouldBe Left(InvalidVatNumber)
       }
 
       "parse a NOT_FOUND response as a VatNumberNotFound" in {
 
-        val mockConnector: GetVatCustomerInformationConnector = new GetVatCustomerInformationConnector(httpClient, mockAppConfig)
-
         val httpResponse = HttpResponse(NOT_FOUND, "")
 
-        val res = mockConnector.GetVatCustomerInformationHttpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
+        val res = httpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
 
         res shouldBe Left(VatNumberNotFound)
       }
 
       "parse a FORBIDDEN response with code MIGRATION as a Migration" in {
 
-        val mockConnector: GetVatCustomerInformationConnector = new GetVatCustomerInformationConnector(httpClient, mockAppConfig)
-
         val httpResponse = HttpResponse(FORBIDDEN, migrationDESJson.toString)
 
-        val res = mockConnector.GetVatCustomerInformationHttpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
+        val res = httpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
 
         res shouldBe Left(Migration)
       }
 
       "parse a FORBIDDEN response without code MIGRATION as a Forbidden" in {
 
-        val mockConnector: GetVatCustomerInformationConnector = new GetVatCustomerInformationConnector(httpClient, mockAppConfig)
-
         val httpResponse = HttpResponse(FORBIDDEN, Json.obj().toString)
 
-        val res = mockConnector.GetVatCustomerInformationHttpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
+        val res = httpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
 
         res shouldBe Left(Forbidden)
       }
 
       "parse any other response as a UnexpectedGetVatCustomerInformationFailure" in {
 
-        val mockConnector: GetVatCustomerInformationConnector = new GetVatCustomerInformationConnector(httpClient, mockAppConfig)
-
         val httpResponse = HttpResponse(INTERNAL_SERVER_ERROR, Json.obj().toString)
 
-        val res = mockConnector.GetVatCustomerInformationHttpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
+        val res = httpParser.GetVatCustomerInformationHttpReads.read(testHttpVerb, testUri, httpResponse)
 
         res shouldBe Left(UnexpectedGetVatCustomerInformationFailure(INTERNAL_SERVER_ERROR, "{}"))
       }
